+++ vulture_whitelist.py
@@ -30,3 +30,6 @@ tz
 model_config
 total_cost
 calculated_total_cost
+
+# @classmethod first parameter (vulture doesn't understand the decorator protocol)
+cls